# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Catalog of indie games — browser or download, free, paid or subscription.",
        "title": "IndieForge API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/games": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Browse games",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search text",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Genre",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "free|paid|subscription|demo",
                        "name": "pricing",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "new|popular|price-asc|price-desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Upload \u0026 publish a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tagline",
                        "name": "tagline",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Genre",
                        "name": "genre",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "free|paid",
                        "name": "pricingModel",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Price (RUB)",
                        "name": "price",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Cover image",
                        "name": "cover",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Screenshots",
                        "name": "screenshots",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "HTML5 build (zip)",
                        "name": "browserBuild",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Downloadable build",
                        "name": "downloadFile",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.GameDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/games/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get a presigned download URL (owners only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game id or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/games/{key}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Game detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game id or slug",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.GameDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/me/games": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Games created by the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "indieforge_internal_dto.DemoDayDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "endsAt": {
                    "type": "string"
                },
                "startsAt": {
                    "type": "string"
                }
            }
        },
        "indieforge_internal_dto.GameDTO": {
            "type": "object",
            "properties": {
                "browserBuildUrl": {
                    "type": "string"
                },
                "canPlayFree": {
                    "type": "boolean"
                },
                "coverImage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "demoDay": {
                    "$ref": "#/definitions/indieforge_internal_dto.DemoDayDTO"
                },
                "description": {
                    "type": "string"
                },
                "developerId": {
                    "type": "string"
                },
                "developerName": {
                    "type": "string"
                },
                "downloadFileName": {
                    "type": "string"
                },
                "downloadPlatforms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "downloadSizeMB": {
                    "type": "integer"
                },
                "friendPackDiscount": {
                    "type": "integer"
                },
                "genre": {
                    "type": "string"
                },
                "hasBrowserBuild": {
                    "type": "boolean"
                },
                "hasDownloadBuild": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "owned": {
                    "type": "boolean"
                },
                "price": {
                    "type": "integer"
                },
                "pricingModel": {
                    "type": "string"
                },
                "screenshots": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "stats": {
                    "$ref": "#/definitions/indieforge_internal_dto.StatsDTO"
                },
                "status": {
                    "type": "string"
                },
                "subscribed": {
                    "type": "boolean"
                },
                "subscription": {
                    "$ref": "#/definitions/indieforge_internal_dto.SubscriptionDTO"
                },
                "supportsMultiplayer": {
                    "type": "boolean"
                },
                "tagline": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "theme": {
                    "$ref": "#/definitions/indieforge_internal_dto.Theme"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "indieforge_internal_dto.StatsDTO": {
            "type": "object",
            "properties": {
                "owners": {
                    "type": "integer"
                },
                "plays": {
                    "type": "integer"
                },
                "subscribers": {
                    "type": "integer"
                }
            }
        },
        "indieforge_internal_dto.SubscriptionDTO": {
            "type": "object",
            "properties": {
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "enabled": {
                    "type": "boolean"
                },
                "period": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "indieforge_internal_dto.Theme": {
            "type": "object",
            "properties": {
                "accent": {
                    "type": "string"
                },
                "accent2": {
                    "type": "string"
                },
                "background": {
                    "type": "string"
                },
                "cardShape": {
                    "type": "string"
                },
                "layout": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}